Support simple folding for preprocessor branches
1. Merge the PR: ##121449 by sr-tream and PR: ##140959 by aketchum15.
2. Fix the issue of `ItaniumMangle.cpp` on the linux platform under clang-9
```c
TypeNameOS << Operand.getKind();
` ` `
The problem of compilation failure caused by the inability to find appropriate operator overloading operations.

+++ clang/lib/AST/ItaniumMangle.cpp
@@ -4729,7 +4729,7 @@ void CXXNameMangler::mangleType(const HLSLInlineSpirvType *T) {
       llvm_unreachable("Invalid SpirvOperand kind");
       break;
     }
-    TypeNameOS << Operand.getKind();
+    TypeNameOS << static_cast<int>(Operand.getKind());
   }
 }
